Time: 0 ms (100.00%) | Memory: 19.6 MB (74.33%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -22,7 +22,7 @@ public:
 
         while (!q.empty()) {
             int levelsize = q.size();
-            int min=q.front().second;
+            long long min=q.front().second;
             
             long long first,last;
